requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.10
botocore==1.38.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.9.0
jmespath==1.0.1
loguru==0.7.3
mcp==1.7.1
openai==1.77.0
pathlib==1.0.1
psutil==7.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.4
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uv==0.7.2
asyncio==3.4.3

# OAuth 2.0 (OIDC) JWT validation dependencies
python-jose[cryptography]==3.3.0
cryptography==42.0.8